requirements.txt
flask==3.0.0
line-bot-sdk==3.11.0
gunicorn==21.2.0
